implement cli and light states

+++ Labs/Milestone/Core/Inc/CLI.h
@@ -1,13 +1,12 @@
 /*
  * CLI.h
  *
- *  Created on: Nov 14, 2024
+ *  Created on: Nov 21, 2024
  *      Author: qkmal
  */
 
 #ifndef INC_CLI_H_
 #define INC_CLI_H_
-
 
 /* Includes ------------------------------------------------------------------*/
 #include "stm32f1xx_hal.h"
@@ -19,12 +18,8 @@ extern uint8_t cliRXChar;
 extern uint8_t cliBufferRX[30];
 extern int counter;
 
-
 //Functions
 uint16_t CLI_Receive();
-
-void CLI_Display(uint16_t);
-
+void CLI_Display(uint8_t);
 
 #endif /* INC_CLI_H_ */
-
